requirements.txt
# Core dependencies
crewai>=0.19.3
crewai-tools>=0.12.0
python-dotenv>=1.0.0
google-genai>=1.40.0
langchain-google-genai>=0.0.6
langchain-core<1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.18.0

# Streamlit
streamlit>=1.31.0

# LLM - Google Gemini support for Agentics library
# google-genai package required for CrewAI Gemini integration

# Utilities
pydantic>=2.6.0
textblob>=0.17.1

# Note: Compatible with Python 3.10 - 3.13
# Tested with Python 3.13.3

